% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams-sp.R
\docType{data}
\name{data-deprecated}
\alias{data-deprecated}
\alias{Pitch_US_Population_2016_v1}
\alias{Pitch_US_Population_2016_v1.states}
\alias{Pitch_US_Population_2016_v1.centers}
\alias{FiveThirtyEightElectoralCollege}
\alias{FiveThirtyEightElectoralCollege.states}
\alias{FiveThirtyEightElectoralCollege.centers}
\alias{NPR1to1}
\alias{NPR1to1.centers}
\alias{DKOS_50_State_OuterHex_Tilemap_v1}
\alias{DKOS_50_State_InnerHex_Tilemap_v1}
\alias{DKOS_50_State_Hex_Tilemap_v1.centers}
\alias{DKOS_CD_Hexmap_v1.1}
\alias{DKOS_CD_Hexmap_v1.1.states}
\alias{DKOS_CD_Hexmap_v1.1.centers}
\alias{DKOS_Electoral_College_Map_v1}
\alias{DKOS_Electoral_College_Map_v1.states}
\alias{DKOS_Electoral_College_Map_v1.centers}
\alias{DKOS_Distorted_Electoral_College_Map_v1}
\alias{DKOS_Distorted_Electoral_College_Map_v1.centers}
\alias{NPR.DemersCartogram}
\alias{NPR.DemersCartogram.centers}
\alias{WP}
\alias{WP.centers}
\alias{WSJ}
\alias{WSJ.centers}
\alias{Datamap.io.tilegram}
\alias{Datamap.io.tilegram.centers}
\title{deprecated sp objects}
\format{See the specific object documentation for the renamed objects.}
\usage{
Pitch_US_Population_2016_v1

Pitch_US_Population_2016_v1.states

Pitch_US_Population_2016_v1.centers

FiveThirtyEightElectoralCollege

FiveThirtyEightElectoralCollege.states

FiveThirtyEightElectoralCollege.centers

NPR1to1

NPR1to1.centers

DKOS_50_State_OuterHex_Tilemap_v1

DKOS_50_State_InnerHex_Tilemap_v1

DKOS_50_State_Hex_Tilemap_v1.centers

DKOS_CD_Hexmap_v1.1

DKOS_CD_Hexmap_v1.1.states

DKOS_CD_Hexmap_v1.1.centers

DKOS_Electoral_College_Map_v1

DKOS_Electoral_College_Map_v1.states

DKOS_Electoral_College_Map_v1.centers

DKOS_Distorted_Electoral_College_Map_v1

DKOS_Distorted_Electoral_College_Map_v1.centers

NPR.DemersCartogram

NPR.DemersCartogram.centers

WP

WP.centers

WSJ

WSJ.centers

Datamap.io.tilegram

Datamap.io.tilegram.centers
}
\description{
These are objects in the older sp format.
  They have been replaced with simple feature (sf) format and will be eventually removed from the package
}
\details{
\code{Pitch_US_Population_2016_v1} is deprecated; please use
  \code{sf_Pitch_US_Population_2016_v1} instead.

\code{Pitch_US_Population_2016_v1.states} is deprecated; please use
  \code{sf_Pitch_US_Population_2016_v1.states} instead.

\code{Pitch_US_Population_2016_v1.centers} is deprecated; please use
  \code{sf_Pitch_US_Population_2016_v1.centers} instead.

\code{FiveThirtyEightElectoralCollege} is deprecated; please use
  \code{sf_FiveThirtyEightElectoralCollege} instead.

\code{FiveThirtyEightElectoralCollege.states} is deprecated; please use
  \code{sf_FiveThirtyEightElectoralCollege.states} instead.

\code{FiveThirtyEightElectoralCollege.centers} is deprecated; please use
  \code{sf_FiveThirtyEightElectoralCollege.centers} instead.

\code{NPR1to1} is deprecated; please use
  \code{sf_NPR1to1} instead.

\code{NPR1to1.centers} is deprecated; please use
  \code{sf_NPR1to1.centers} instead.

\code{DKOS_50_State_OuterHex_Tilemap_v1} is deprecated; please use
  \code{sf_DKOS_50_State_OuterHex_Tilemap_v1} instead.

\code{DKOS_50_State_InnerHex_Tilemap_v1} is deprecated; please use
  \code{sf_DKOS_50_State_InnerHex_Tilemap_v1} instead.

\code{DKOS_50_State_Hex_Tilemap_v1.centers} is deprecated; please use
  \code{sf_DKOS_50_State_Hex_Tilemap_v1.centers} instead.

\code{DKOS_CD_Hexmap_v1.1} is deprecated; please use
  \code{sf_DKOS_CD_Hexmap_v1.1} instead.

\code{DKOS_CD_Hexmap_v1.1.states} is deprecated; please use
  \code{sf_DKOS_CD_Hexmap_v1.1.states} instead.

\code{DKOS_CD_Hexmap_v1.1.centers} is deprecated; please use
  \code{sf_DKOS_CD_Hexmap_v1.1.centers} instead.

\code{DKOS_Electoral_College_Map_v1} is deprecated; please use
  \code{sf_DKOS_Electoral_College_Map_v1} instead.

\code{DKOS_Electoral_College_Map_v1.states} is deprecated; please use
  \code{sf_DKOS_Electoral_College_Map_v1.states} instead.

\code{DKOS_Electoral_College_Map_v1.centers} is deprecated; please use
  \code{sf_DKOS_Electoral_College_Map_v1.centers} instead.

\code{DKOS_Distorted_Electoral_College_Map_v1} is deprecated; please use
  \code{sf_DKOS_Distorted_Electoral_College_Map_v1} instead.

\code{DKOS_Distorted_Electoral_College_Map_v1.centers} is deprecated; please use
  \code{sf_DKOS_Distorted_Electoral_College_Map_v1.centers} instead.

\code{NPR.DemersCartogram} is deprecated; please use
  \code{sf_NPR.DemersCartogram} instead.

\code{NPR.DemersCartogram.centers} is deprecated; please use
  \code{sf_NPR.DemersCartogram.centers} instead.

\code{WP} is deprecated; please use
  \code{sf_WP} instead.

\code{WP.centers} is deprecated; please use
  \code{sf_WP.centers} instead.

\code{WSJ} is deprecated; please use
  \code{sf_WSJ} instead.

\code{WSJ.centers} is deprecated; please use
  \code{sf_WSJ.centers} instead.

\code{Datamap.io.tilegram} is deprecated; please use
  \code{sf_Datamap.io.tilegram} instead.

\code{Datamap.io.tilegram.centers} is deprecated; please use
  \code{sf_Datamap.io.tilegram.centers} instead.
}
\keyword{datasets}
\keyword{deprecated}
\keyword{internal}
