% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{Datamap.io.tilegram}
\alias{Datamap.io.tilegram}
\alias{Datamap.io.tilegram.centers}
\title{Datamap.io.tilegram}
\format{A SpatialPolygonsDataFrame}
\usage{
Datamap.io.tilegram

Datamap.io.tilegram.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- Datamap.io.tilegram
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

