% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{DKOS_50_State_OuterHex_Tilemap_v1}
\alias{DKOS_50_State_Hex_Tilemap_v1.centers}
\alias{DKOS_50_State_InnerHex_Tilemap_v1}
\alias{DKOS_50_State_OuterHex_Tilemap_v1}
\title{DKOS_50_State_OuterHex_Tilemap_v1}
\format{A SpatialPolygonsDataFrame}
\usage{
DKOS_50_State_OuterHex_Tilemap_v1

DKOS_50_State_InnerHex_Tilemap_v1

DKOS_50_State_Hex_Tilemap_v1.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\keyword{datasets}

