% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{DKOS_CD_Hexmap_v1.1}
\alias{DKOS_CD_Hexmap_v1.1}
\alias{DKOS_CD_Hexmap_v1.1.centers}
\alias{DKOS_CD_Hexmap_v1.1.states}
\title{DKOS_CD_Hexmap_v1.1}
\format{A SpatialPolygonsDataFrame}
\usage{
DKOS_CD_Hexmap_v1.1

DKOS_CD_Hexmap_v1.1.states

DKOS_CD_Hexmap_v1.1.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon equals one congressional district

A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- DKOS_CD_Hexmap_v1.1
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

