% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{DKOS_Distorted_Electoral_College_Map_v1}
\alias{DKOS_Distorted_Electoral_College_Map_v1}
\alias{DKOS_Distorted_Electoral_College_Map_v1.centers}
\title{DKOS_Distorted_Electoral_College_Map_v1}
\format{A SpatialPolygonsDataFrame}
\usage{
DKOS_Distorted_Electoral_College_Map_v1

DKOS_Distorted_Electoral_College_Map_v1.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- DKOS_Distorted_Electoral_College_Map_v1
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

