% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{FiveThirtyEightElectoralCollege.centers}
\alias{FiveThirtyEightElectoralCollege.centers}
\title{FiveThirtyEightElectoralCollege.centers}
\format{A SpatialPointsDataFrame}
\usage{
FiveThirtyEightElectoralCollege.centers
}
\description{
A `SpatialPointsDataFrame`  of centroids of each state
}
\keyword{datasets}

