% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{FiveThirtyEightElectoralCollege}
\alias{FiveThirtyEightElectoralCollege}
\title{FiveThirtyEightElectoralCollege}
\format{A SpatialPolygonsDataFrame}
\usage{
FiveThirtyEightElectoralCollege
}
\description{
A `SpatialPolygonsDataFrame` where each polygon = 500K people
}
\keyword{datasets}

