% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{FiveThirtyEightElectoralCollege.states}
\alias{FiveThirtyEightElectoralCollege.states}
\title{FiveThirtyEightElectoralCollege.states}
\format{A SpatialPolygonsDataFrame}
\usage{
FiveThirtyEightElectoralCollege.states
}
\description{
A `SpatialPolygonsDataFrame` where each polygon = 1 state
}
\keyword{datasets}

