% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{NPR.DemersCartogram}
\alias{NPR.DemersCartogram}
\alias{NPR.DemersCartogram.centers}
\title{NPR.DemersCartogram}
\format{A SpatialPolygonsDataFrame}
\usage{
NPR.DemersCartogram

NPR.DemersCartogram.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- NPR.DemersCartogram
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

