% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{NPR.DorlingCartogram}
\alias{NPR.DorlingCartogram}
\alias{NPR.DorlingCartogram.centers}
\title{NPR.DorlingCartogram}
\format{A SpatialPolygonsDataFrame}
\usage{
NPR.DorlingCartogram

NPR.DorlingCartogram.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\keyword{datasets}

