% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{NPR1to1.centers}
\alias{NPR1to1.centers}
\title{NPR1to1.centers}
\format{A SpatialPointsDataFrame}
\usage{
NPR1to1.centers
}
\description{
A `SpatialPointsDataFrame`  of centroids of each state
}
\keyword{datasets}

