% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{NPR1to1}
\alias{NPR1to1}
\title{NPR1to1}
\format{A SpatialPolygonsDataFrame}
\usage{
NPR1to1
}
\description{
A `SpatialPolygonsDataFrame` where each polygon = 500K people
}
\keyword{datasets}

