% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{Pitch_US_Population_2016_v1.centers}
\alias{Pitch_US_Population_2016_v1.centers}
\title{Pitch_US_Population_2016_v1.centers}
\format{A SpatialPointsDataFrame}
\usage{
Pitch_US_Population_2016_v1.centers
}
\description{
A `SpatialPointsDataFrame`  of centroids of each state
}
\keyword{datasets}

