% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{Pitch_US_Population_2016_v1}
\alias{Pitch_US_Population_2016_v1}
\title{Pitch_US_Population_2016_v1}
\format{A SpatialPolygonsDataFrame}
\usage{
Pitch_US_Population_2016_v1
}
\description{
A `SpatialPolygonsDataFrame` where each polygon = 500K people
}
\keyword{datasets}

