% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{Pitch_US_Population_2016_v1.states}
\alias{Pitch_US_Population_2016_v1.states}
\title{Pitch_US_Population_2016_v1.states}
\format{A SpatialPolygonsDataFrame}
\usage{
Pitch_US_Population_2016_v1.states
}
\description{
A `SpatialPolygonsDataFrame` where each polygon = 1 state
}
\keyword{datasets}

