% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_Datamap.io.tilegram}
\alias{sf_Datamap.io.tilegram}
\alias{sf_Datamap.io.tilegram.centers}
\title{sf_Datamap.io.tilegram}
\format{sf}
\usage{
sf_Datamap.io.tilegram

sf_Datamap.io.tilegram.centers
}
\description{
A `sf` object where each polygon represents a state

A `sf` object of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
data <- sf_Datamap.io.tilegram
leaflet(data,
  options=leafletOptions(crs=leafletCRS("L.CRS.Simple"))) \%>\%
  addPolygons()
}
}
\keyword{datasets}
