% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_DKOS_50_State_OuterHex_Tilemap_v1}
\alias{sf_DKOS_50_State_OuterHex_Tilemap_v1}
\alias{sf_DKOS_50_State_InnerHex_Tilemap_v1}
\alias{sf_DKOS_50_State_Hex_Tilemap_v1.centers}
\title{sf_DKOS_50_State_OuterHex_Tilemap_v1}
\format{sf}
\usage{
sf_DKOS_50_State_OuterHex_Tilemap_v1

sf_DKOS_50_State_InnerHex_Tilemap_v1

sf_DKOS_50_State_Hex_Tilemap_v1.centers
}
\description{
A `sf` object where each polygon represents a state

A `sf` object where each polygon represents a state

A `sf` object of centroids of each state
}
\keyword{datasets}
