% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_DKOS_Electoral_College_Map_v1}
\alias{sf_DKOS_Electoral_College_Map_v1}
\alias{sf_DKOS_Electoral_College_Map_v1.states}
\alias{sf_DKOS_Electoral_College_Map_v1.centers}
\title{sf_DKOS_Electoral_College_Map_v1}
\format{sf}
\usage{
sf_DKOS_Electoral_College_Map_v1

sf_DKOS_Electoral_College_Map_v1.states

sf_DKOS_Electoral_College_Map_v1.centers
}
\description{
A `sf` object where each polygon equals one electoral college

A `sf` object where each polygon represents a state

A `sf` object of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
data <- sf_DKOS_Electoral_College_Map_v1
leaflet(data,
  options=leafletOptions(crs=leafletCRS("L.CRS.Simple"))) \%>\%
  addPolygons()
}
}
\keyword{datasets}
