% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_france_all_regions_population}
\alias{sf_france_all_regions_population}
\alias{sf_france_all_regions_population.regions}
\alias{sf_france_all_regions_population.centers}
\title{sf_france_all_regions_population}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 18 rows and 4 columns.}
\usage{
sf_france_all_regions_population

sf_france_all_regions_population.regions

sf_france_all_regions_population.centers
}
\description{
A `sf` object for regions of France including overseas.

A `sf` object of regional boundaries of each French region.

A `sf` object of centroids of each French region.
}
\keyword{datasets}
