% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_france_departments}
\alias{sf_france_departments}
\alias{sf_france_departments.centers}
\title{sf_france_departments}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 96 rows and 4 columns.}
\usage{
sf_france_departments

sf_france_departments.centers
}
\description{
A `sf` object for departments of France.

A `sf` object of centroids of each French department.
}
\keyword{datasets}
