% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_germany_constituencies}
\alias{sf_germany_constituencies}
\alias{sf_germany_constituencies.centers}
\title{sf_germany_constituencies}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 299 rows and 4 columns.}
\usage{
sf_germany_constituencies

sf_germany_constituencies.centers
}
\description{
A `sf` object for regions of Germany

A `sf` object of centroids of each German region.
}
\keyword{datasets}
