% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_NPR.DemersCartogram}
\alias{sf_NPR.DemersCartogram}
\alias{sf_NPR.DemersCartogram.centers}
\title{sf_NPR.DemersCartogram}
\format{sf}
\usage{
sf_NPR.DemersCartogram

sf_NPR.DemersCartogram.centers
}
\description{
A `sf` object where each polygon represents a state

A `sf` object of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
data <- sf_NPR.DemersCartogram
leaflet(data,
  options=leafletOptions(crs=leafletCRS("L.CRS.Simple"))) \%>\%
  addPolygons()
}
}
\keyword{datasets}
