% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{sf_WSJ}
\alias{sf_WSJ}
\alias{sf_WSJ.centers}
\title{sf_WSJ}
\format{sf}
\usage{
sf_WSJ

sf_WSJ.centers
}
\description{
A `sf` object where each polygon represents a state

A `sf` object of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- sf_WSJ
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}
