% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegramsR-package.r
\docType{package}
\name{tilegramsR}
\alias{tilegramsR}
\alias{tilegramsR-package}
\title{tilegramsR}
\description{
Tilegrams are tiled maps where the region size is proportional to
 the certain characteristics of the dataset. This package provides several
 such tilegrams as simple feature (sf) objects.
}
