% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{WP}
\alias{WP}
\alias{WP.centers}
\title{WP}
\format{A SpatialPolygonsDataFrame}
\usage{
WP

WP.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- WP
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

