% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{WSJ}
\alias{WSJ}
\alias{WSJ.centers}
\title{WSJ}
\format{A SpatialPolygonsDataFrame}
\usage{
WSJ

WSJ.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- WSJ
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

