% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{FiveThirtyEightElectoralCollege}
\alias{FiveThirtyEightElectoralCollege}
\alias{FiveThirtyEightElectoralCollege.centers}
\alias{FiveThirtyEightElectoralCollege.states}
\title{FiveThirtyEightElectoralCollege}
\format{A SpatialPolygonsDataFrame}
\usage{
FiveThirtyEightElectoralCollege

FiveThirtyEightElectoralCollege.states

FiveThirtyEightElectoralCollege.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon = 500K people

A `SpatialPolygonsDataFrame` where each polygon = 1 state

A `SpatialPointsDataFrame`  of centroids of each state
}
\keyword{datasets}

