% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{NPR1to1}
\alias{NPR1to1}
\alias{NPR1to1.centers}
\title{NPR1to1}
\format{A SpatialPolygonsDataFrame}
\usage{
NPR1to1

NPR1to1.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- NPR1to1
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

