% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilegrams.R
\docType{data}
\name{Pitch_US_Population_2016_v1}
\alias{Pitch_US_Population_2016_v1}
\alias{Pitch_US_Population_2016_v1.centers}
\alias{Pitch_US_Population_2016_v1.states}
\title{Pitch_US_Population_2016_v1}
\format{A SpatialPolygonsDataFrame}
\usage{
Pitch_US_Population_2016_v1

Pitch_US_Population_2016_v1.states

Pitch_US_Population_2016_v1.centers
}
\description{
A `SpatialPolygonsDataFrame` where each polygon equals 500K people

A `SpatialPolygonsDataFrame` where each polygon represents a state

A `SpatialPointsDataFrame` of centroids of each state
}
\examples{
\dontrun{
library(leaflet)
library(tilegramsR)
spdf <- Pitch_US_Population_2016_v1
leaflet(spdf) \%>\% addPolygons()
}
}
\keyword{datasets}

